# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CourseOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "global_dataset_name": {
            "title": "Global Dataset Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "term": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "term",
          "owner_user_id",
          "global_dataset_name",
          "created_at"
        ],
        "title": "CourseOut",
        "type": "object"
      },
      "CourseRef": {
        "properties": {
          "course": {
            "pattern": "^[a-z][a-z0-9]{1,15}$",
            "title": "Course",
            "type": "string"
          }
        },
        "required": [
          "course"
        ],
        "title": "CourseRef",
        "type": "object"
      },
      "CourseSearchHit": {
        "properties": {
          "course": {
            "$ref": "#/components/schemas/CourseOut"
          },
          "enrolled": {
            "title": "Enrolled",
            "type": "boolean"
          }
        },
        "required": [
          "course",
          "enrolled"
        ],
        "title": "CourseSearchHit",
        "type": "object"
      },
      "CourseSearchOut": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/CourseSearchHit"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "CourseSearchOut",
        "type": "object"
      },
      "CreateCourse": {
        "properties": {
          "code": {
            "pattern": "^[a-z][a-z0-9]{1,15}$",
            "title": "Code",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "term": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "CreateCourse",
        "type": "object"
      },
      "EnrolmentOut": {
        "properties": {
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "user_dataset_name": {
            "title": "User Dataset Name",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "course_id",
          "user_dataset_name",
          "created_at"
        ],
        "title": "EnrolmentOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateCourse": {
        "properties": {
          "course": {
            "pattern": "^[a-z][a-z0-9]{1,15}$",
            "title": "Course",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "term": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          }
        },
        "required": [
          "course"
        ],
        "title": "UpdateCourse",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes_opt_out": {
            "title": "Notes Opt Out",
            "type": "boolean"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "notes_opt_out",
          "created_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Lattice API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/courses.create": {
      "post": {
        "description": "One course per code. An existing code is a 409 carrying it, so the caller can join.",
        "operationId": "create_course_courses_create_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses.get": {
      "get": {
        "operationId": "get_course_courses_get_get",
        "parameters": [
          {
            "in": "query",
            "name": "course",
            "required": true,
            "schema": {
              "title": "Course",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses.list": {
      "get": {
        "operationId": "list_courses_courses_list_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CourseOut"
                  },
                  "title": "Response List Courses Courses List Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Courses",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses.search": {
      "get": {
        "description": "Exact code or substring title, with the caller's enrolment state on each hit.",
        "operationId": "search_courses_courses_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseSearchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Courses",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses.update": {
      "post": {
        "operationId": "update_course_courses_update_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/enrolments.join": {
      "post": {
        "description": "Record the enrolment and create the engine-side datasets it grants.",
        "operationId": "join_course_enrolments_join_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRef"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrolmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Join Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/enrolments.leave": {
      "post": {
        "description": "Drops the record only; the private Dataset and its Notes survive a re-join.",
        "operationId": "leave_course_enrolments_leave_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRef"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Leave Course Enrolments Leave Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Leave Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/enrolments.list": {
      "get": {
        "operationId": "list_enrolments_enrolments_list_get",
        "parameters": [
          {
            "in": "query",
            "name": "course",
            "required": true,
            "schema": {
              "title": "Course",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "title": "Response List Enrolments Enrolments List Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Enrolments",
        "tags": [
          "courses"
        ]
      }
    },
...
  }
}